Added a setting to control if video tracks are ignored
Users may want to keep tracking video tracks, for example if they rip
the audio part of the video tracks. Added a setting to allow this. This
is related to the #2688 change.

My impression is that it would be quite rare though so I made the
setting ignore video tracks by default, but that's easily changed if
desired.